Split test_proxyfs into external files. NFC
This was the longest test in `test_other.py` weighing it at 302 lines.
