+++ test/other/test_proxyfs_main.js
@@ -1,0 +1,96 @@
+var m0 = require('./test_proxyfs.js');
+var m1 = require('./test_proxyfs1.js');
+var m2 = require('./test_proxyfs2.js');
+
+var section;
+function print(str){
+  process.stdout.write(section+":"+str+":");
+}
+
+m0.FS.mkdir('/working');
+m0.FS.mount(m0.PROXYFS,{root:'/',fs:m1.FS},'/working');
+m0.FS.mkdir('/working2');
+m0.FS.mount(m0.PROXYFS,{root:'/',fs:m2.FS},'/working2');
+
+section = "child m1 reads and writes local file.";
+print("m1 read embed");
+m1.ccall('myreade','number',[],[]);
+print("m1 write");console.log("");
+m1.ccall('mywrite0','number',['number'],[1]);
+print("m1 read");
+m1.ccall('myread0','number',[],[]);
+
+section = "child m2 reads and writes local file.";
+print("m2 read embed");
+m2.ccall('myreade','number',[],[]);
+print("m2 write");console.log("");
+m2.ccall('mywrite0','number',['number'],[2]);
+print("m2 read");
+m2.ccall('myread0','number',[],[]);
+
+section = "child m1 reads local file.";
+print("m1 read");
+m1.ccall('myread0','number',[],[]);
+
+section = "parent m0 accesses children's file.";
+print("m0 access existing");
+m0.ccall('myaccess0existing','number',[],[]);
+print("m0 access absent");
+m0.ccall('myaccess0absent','number',[],[]);
+
+section = "child m1 accesses local file.";
+print("m1 access existing");
+m1.ccall('myaccess1existing','number',[],[]);
+print("m1 access absent");
+m1.ccall('myaccess1absent','number',[],[]);
+
+section = "parent m0 reads and writes local and children's file.";
+print("m0 read embed");
+m0.ccall('myreade','number',[],[]);
+print("m0 read m1");
+m0.ccall('myread1','number',[],[]);
+print("m0 read m2");
+m0.ccall('myread2','number',[],[]);
+
+section = "m0,m1 and m2 verify local files.";
+print("m0 write");console.log("");
+m0.ccall('mywrite0','number',['number'],[0]);
+print("m0 read");
+m0.ccall('myread0','number',[],[]);
+print("m1 read");
+m1.ccall('myread0','number',[],[]);
+print("m2 read");
+m2.ccall('myread0','number',[],[]);
+
+print("m0 read embed");
+m0.ccall('myreade','number',[],[]);
+print("m1 read embed");
+m1.ccall('myreade','number',[],[]);
+print("m2 read embed");
+m2.ccall('myreade','number',[],[]);
+
+section = "parent m0 writes and reads children's files.";
+print("m0 write m1");console.log("");
+m0.ccall('mywrite1','number',[],[]);
+print("m0 read m1");
+m0.ccall('myread1','number',[],[]);
+print("m0 write m2");console.log("");
+m0.ccall('mywrite2','number',[],[]);
+print("m0 read m2");
+m0.ccall('myread2','number',[],[]);
+print("m1 read");
+m1.ccall('myread0','number',[],[]);
+print("m2 read");
+m2.ccall('myread0','number',[],[]);
+print("m0 read m0");
+m0.ccall('myread0','number',[],[]);
+
+section = "parent m0 renames a file in child fs.";
+m0.FS.writeFile('/working/test', 'testme');
+m0.FS.rename('/working/test', '/working/test.bak');
+console.log(section + ":renamed file accessible by the new name:" + m0.FS.analyzePath('/working/test.bak').exists);
+console.log(section + ":renamed file accessible by the old name:" + m0.FS.analyzePath('/working/test').exists);
+
+section = "test seek.";
+print("file size");
+m0.ccall('myreadSeekEnd', 'number', [], []);
